% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\docType{methods}
\name{ConvertirEnTexto-Method}
\alias{ConvertirEnTexto-Method}
\alias{ConvertirEnTexto,integer-method}
\alias{ConvertirEnTexto,numeric-method}
\title{Convertir a texto}
\usage{
\S4method{ConvertirEnTexto}{integer}(text)

\S4method{ConvertirEnTexto}{numeric}(text)
}
\arguments{
\item{text}{El valor que será convertido en texto}
}
\description{
Descripción: Esta función convierte un entero a un objeto de texto.

Descripción: Esta función convierte un número a un objeto de texto.
}
\details{
Más detalles aquí.

y aquí...
}
\section{Functions}{
\itemize{
\item \code{ConvertirEnTexto,integer-method}: 

\item \code{ConvertirEnTexto,numeric-method}: 
}}

\section{Una sección personalizada}{


Detalles de la sección personalizada.
}

\examples{
ConvertirEnTexto(3) ## returns "3"

## don't run this in calls to 'example(add_numbers)'
\dontrun{
   ConvertirEnTexto(3) ## returns "3"
}

## don't test this during 'R CMD check'
\donttest{
   ConvertirEnTexto(3) ## returns "3"
}
ConvertirEnTexto(3.0) ## returns "3"

## don't run this in calls to 'example(add_numbers)'
\dontrun{
   ConvertirEnTexto(3.0) ## returns "3"
}

## don't test this during 'R CMD check'
\donttest{
   ConvertirEnTexto(3.0) ## returns "3"
}
}
