% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aGeneric.R
\name{ConvertirEnTexto}
\alias{ConvertirEnTexto}
\title{Genéricos}
\usage{
ConvertirEnTexto(text)
}
\value{
Un objeto de tipo definido en \code{valueClass}. En este caso, devuelve un texto.
}
\description{
Los genéricos proveen el nombre del método, pero no la definición.
}
\details{
Usar un genérico para verbos comunes como 'Escribir', 'Borrar', 'Extraer', 'Guardar'. La acción es la misma sobre cualquier objeto, pero los detalles de implementación se definen en \code{setMethod}.

En este ejemplo, la función convierte números reales y enteros a texto. Observar que el argumento \code{signature} solo acepta ese tipo de objetos; de lo contrario, da error.
}
