% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clases.R
\docType{class}
\name{Persona-class}
\alias{Persona-class}
\title{Una persona}
\description{
Una persona
}
\section{Slots}{

\describe{
\item{\code{nombre}}{El nombre de la persona}

\item{\code{edad}}{La edad de la persona}
}}

