% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clases.R
\docType{methods}
\name{Persona-Method}
\alias{Persona-Method}
\alias{Cumplir,Persona,integer-method}
\title{Métodos para persona}
\usage{
\S4method{Cumplir}{Persona,integer}(persona, cuantos)
}
\arguments{
\item{cuantos}{La edad deseada para la persona.}
}
\value{
Una persona
}
\description{
Descripción: Incrementa la edad de la persona.
}
\section{Functions}{
\itemize{
\item \code{Cumplir,Persona,integer-method}: 
}}

